Give script upload form its own file
